use jabber, not the API, to send issue messages
Unfortunately, we can't use the API to respond to private messages,
and we can't send nicely formatted messages via jabber. In the issue
case, we need private messages, so we need to use jabber. This means
I can't send good looking messages.

I've contacted hipchat support to ask them to improve the situation.