namespace the connector and Message classes
There's a new transitive dependency on the console gem which conflicts
with our Console connector at the root of the module/class namespace.
So I wrapped our connectors in a Linkbot module. Confirmed that Console
and Slack work which are the ones we mostly care about these days.

Wrapping Message and MessageType in the Linkbot module sorted out the
connectors being able to use Message within the namespace. Also,
probably avoids some future naming conflict with a message gem that
sneaks its way into dependencies.