update test run to include the specs for individual plugins
Specs for specific plugins live outside the ./spec directory, so add the
directory to the rake task.

Running the plugin spec and the specs of plugins threw off the
expectation that there would be only one plugin in the collection, so
updated the plugins_spec to be more selective about what it is testing.

Removed the unused spec_helper and a few vestigial lines from the real
helper.